rename JsonDataProvider to JsonDataHandler
