+++ src/main/java/com/safetynet/safetynetalertsapi/repositories/DataHandler.java
@@ -1,0 +1,9 @@
+package com.safetynet.safetynetalertsapi.repositories;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.safetynet.safetynetalertsapi.model.DataSet;
+
+public interface DataHandler {
+	public DataSet getAllData() throws JsonProcessingException;
+
+}
