+++ src/main/java/com/safetynet/safetynetalertsapi/services/validators/FireStationValidator.java
@@ -6,16 +6,16 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.safetynet.safetynetalertsapi.model.FireStation;
-import com.safetynet.safetynetalertsapi.repositories.JsonDataProvider;
+import com.safetynet.safetynetalertsapi.repositories.JsonDataHandler;
 
 @Service
 public class FireStationValidator {
 	
 	@Autowired
-	JsonDataProvider provider;
+	JsonDataHandler dataHandler;
 	
 	public boolean stationExists(int stationNumber) {
-		List<FireStation> fireStations = provider.getAllData().getFireStations();
+		List<FireStation> fireStations = dataHandler.getAllData().getFireStations();
 		
 		boolean exists = fireStations.stream()
 		.anyMatch(station -> station.getStation() == stationNumber);
